Lots of additions and changes.
??-Jul-2017 Yukiko Changed  Made some minor fixes, changes, and updates
to some files but I don't remember when or what was changed. It happens.

22-Jul-2017 Yukiko Added  The "decorate the facets" package to the
Distro. Also added was the clilocs package because it is needed by the
facet decoration package. It also might prove useful later in future
work on the Distro.

23-Jul-2017 Yukiko Updated  The doors itemdesc file to include new doors
that weren't in the file including sliding doors. I also added addition
single tile book shelves to the carpentry itemdesc file.

24-Jul-2017 Yukiko Changed  The way book spawning shelves are handled.
They are now handled by a Control Script. Prior to this a Global
property was set for every shelf that spawned books. This property
became quite large in pol.txt which seemed ridiculous. So now the
control script handles this. There are 2 CProps that get placed on a
book shelf that is to spawn books, Spawn and MaxBooks. Spawn is set to 1
and MaxBooks any integer from 1 to 15. The control script will check
every 10 minutes and spawn the necessary number of books to achieve
MaxBooks. I also added a command, spawnShelf, that will allow GMs to
turn any non-spawning book shelf that graphically looks like it contains
books into a book spawning shelf.

24-Jul-2017 Yukiko Changed  The handling of damage in various scripts to
now call the damage functions in damage.inc which take into account
resistances such as Fire, Clod etcetera.

25-Jun-2017 Yukiko Changed  Removed the prefix "ATTRIBUTEID_" from the
skill attributes and changed the names of some like EVALINT and ITEMID
to  longer names, EVALUATING_INTELLIGENCE and ITEM_IDENTIFICATION. See
\scripts\include\attributes.inc for the new naming convention. Updated
the skills to include Bushido, Ninjitsu, Spellweaving,Mysticism,
Imbuing, and Throwing bringing the skill count to 57 snd in line with
ToL. Those skills aren't usable yet but I'm working on it.

25-Jul-2017 Yukiko Added  The Necromancy scrolls defined in the client
and incorporated the Necromancy spells from Arena PVP shard. I modified
some of the spells to comply with info provided on UO Guide wiki but did
not get to all of them. Arena shard didn't apparently add in all the
extra stuff the spells are supposed to do. I can't be certain about that
though without a greater study of their other scripts. I also added a
couple of NPCs necessary to the necromancy package.

25-Jul-2017 Yukiko Added  The Necromancy scrolls to nlootgroup.

25-Jul-2017 Yukiko Added  The way resistances are handled by equip and
unequip scrpts. Damage.inc does not use the resistance members to handle
resistances but rather places a CProp on the mobile which contains a
dictionary of resistances. I added a check for resistance members on
items and then passed the value to ModifyResistance in damage.inc