Added runic atlas and new rune book and some other changes.
From the Distro changes.txt:
31-Jul-2017 Yukiko Added  Added the Runic Atlas to the Distro
contributed by ThisIsMe. Also added itemUtil.inc from ThisIsMe's shard.
It's required for the Runic Atlas and I presume for the Rune Book that
ThisIsMe is redesigning and will most likely be added to the Distro.
I'll have to decide how to handle that if it is added so that old style
rune books aren't broken, just in case they exist in someone's server.

31-Jul-2017 Yukiko Changed  Previously rune destination info was stored
as individual CProps for x, y, z, realm destinations. I changed it to a
struct CProp containing the info. I also moved recall runes from the
main itemdesc file to a package "runes" and gave them methods. One of
the methods is GetDestination(). In that method I added a check for the
old destination CProps. If they are present the method will build the
struct for the new way of marking the runes, save the struct CProp on
the rune and remove the deprecated CProps. This will ensure any runes
created prior to the change will be functional and get updated in the
process.

2-Aug-2017 Yukiko Added  A reworked rune book contributed by ThisIsMe.
The old rune books were left in to prevent a server shut down just in
case you have them lying around but they currently do not work properly
and should be replaced with the current rune books. the old rune books
item name is oldrunebook but it doesn't open the correct gump pic for
some reason so get rid of them. Sorry for the abrupt deprecation of an
item. I strive for legacy support but in this case something broke and I
haven't had time to look into it.

6-Aug-2017 Yukiko Fixed  The book spawn system was causing excess server
load. So DevGIB and I worked out a fix (hopefully) that will solve this.
Also cleaned up some code from the old book spawn system.