Corrected an error
I mitakenly uploaded an interim math.inc file and accidentally replace
math.inc version 2.0 with version 1.3. This commit corrects that error.
The new math.inc version is 2.1 with the addition of ThisIsMe's
functions that return "dice strings" based on min and max values. These
were in the 1.3 version as well.