Restructured attributes.inc location.
I have moved the current version of attributes.inc and the associated
cfg files from their former locations to the attributes pkg. I changed
all current scripts which used attributess.inc which used to be in
\scripts\include to reference that in the attributes pkg. A copy of
attributes.inc still exists in \scripts\include for compatability but
all new submissions should reference it in the attributes pkg.