Damage.inc changrd to use resistance members.
The previous version of damage.inc used a dictionary CProp to store resistance values for the 5 elemental resistances. Now it reads the members froim the mobile to get resistance values. Fixed spellStarter.src.. Changed  the name of the damage type member of the struct it sends to spell scripts from damage_typ to damage_type. The spell scripts were looking for the name spelled as it is now. Easier fixing one script than al the spell damage scripts.